feat: Fix edge cases where transforms look weird due to jump scrolling
Essentially add a quick fix where if we pass a section
completely we make sure the transform we want stays e.g.
for the parallax, once the parallax section comes in view,
we want the brightness to stay as 1 and not to increase as
you continue scrolling.

Fixed by adding a condition that checks if we've scrolled
the section into view. If we have, we max the max scroll
or max scrollDistance so we maintain the same transform.