feat: Finished project MVP (or like 90%)
Only thing left is to add the extra up and down
transitions from moving the text elements up/down
more when they are leaving.

Gonna choose to leave it as something ik how to do
in my head kinda thing (don't think i'll work on this
anymore).

Not until I watch the vid on how to scroll animate a
video.